Overwrite configure.py-modified __init__.py with original
On Windows, __init__.py is modified by configure.py to find the first
directory on the user's PATH that includes Qt5Core.dll and add it to the
DLL search paths with os.add_dll_directory(). Conda-forge's qt-main does
not distribute Qt5Core.dll, instead calling it Qt5Core_conda.dll, so
this modification only serves to allow for external Qt installations to
interfere with the conda one in certain situations. This removes that
modification since it is at best not needed and at worst introduces
bugs.