Overwrite configure.py-modified __init__.py with original
On Windows, __init__.py is modified by configure.py to find the first
directory on the user's PATH that includes Qt6Core.dll and add it to the
DLL search paths with os.add_dll_directory(). Though unlike qt-main
Conda-forge's qt6-main does keep the name Qt6Core.dll and so this search
would probably work out fine, this modification of the search path could
still allow for external Qt installations to interfere with the conda
one in certain situations. This removes that modification since it is at
best not needed and at worst introduces bugs.